requirements.txt
tiktoken==0.4.0
chromadb==0.4.5
openai
langchain
python-dotenv
pinecone-client